+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release 5.1.0:
  - Drop ARIES protocol implementation, and the `vck-aries` artifact
  - Add `credentialScheme` and `subjectPublicKey` to internal `CredentialToBeIssued`
  - Refactor `issueCredential` of `Issuer` to directly get the credential-to-be-issued
+ - Replace `buildIssuerCredentialDataProviderOverride` in `CredentialIssuer` with `credentialProvider` to extract user information into a credential
 
 Release 5.0.1:
  - Update JsonPath4K to 2.4.0